requirements.txt
streamlit>=1.24
openai>=1.0.0
python-dotenv>=1.0.0
extra-streamlit-components>=0.1.70
